+++ src-examples/BuilderConsumer/Settings.cs
@@ -4,6 +4,6 @@ namespace BuilderConsumer
 {
     public class Settings
     {
-        public IReadOnlyCollection<Content> Contents { get; set; }
+        public IReadOnlyCollection<Content> Contents { get; set; } = [];
     }
 }