+++ src-examples/BuilderConsumerNET8_ViaNuGet/ClassWithPrivateSetter.cs
@@ -1,6 +1,6 @@
 using FluentBuilder;
 
-namespace BuilderConsumerNET6
+namespace BuilderConsumerNET8
 {
     [AutoGenerateBuilder(FluentBuilderAccessibility.Public)]
     public class ClassWithPrivateSetter
@@ -8,5 +8,11 @@ namespace BuilderConsumerNET6
         public int Value1 { get; private set; }
 
         public int Value2 { get; set; }
+
+        [AutoGenerateBuilder(FluentBuilderAccessibility.PublicAndPrivate)]
+        private class PrivateClass
+        {
+            public int Test { get; set; }
+        }
     }
 }