+++ src-examples/BuilderConsumerNET8_ViaNuGet/Option.cs
@@ -1,4 +1,4 @@
-namespace BuilderConsumerNET6;
+namespace BuilderConsumerNET8;
 
 public class Option
 {