+++ src-extensions/CSharp.SourceGenerators.Extensions/Compatibility/RequiredMemberAttribute.cs
@@ -1,0 +1,13 @@
+// ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices;
+
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
+internal sealed class RequiredMemberAttribute : Attribute;
+
+[AttributeUsage(AttributeTargets.All, AllowMultiple = true, Inherited = false)]
+internal sealed class CompilerFeatureRequiredAttribute(string featureName) : Attribute
+{
+    public string FeatureName { get; } = featureName;
+
+    public bool IsOptional { get; init; }
+}