+++ src/FluentBuilderGenerator/FileGenerators/FluentBuilderClassesGenerator.IDictionary.cs
@@ -1,5 +1,6 @@
 using System.Text;
 using FluentBuilderGenerator.Extensions;
+using FluentBuilderGenerator.Interfaces;
 using FluentBuilderGenerator.Models;
 using Microsoft.CodeAnalysis;
 
@@ -9,7 +10,7 @@ internal partial class FluentBuilderClassesGenerator
 {
     private StringBuilder GenerateWithIDictionaryBuilderActionMethod(
         ClassSymbol classSymbol,
-        IPropertySymbol property,
+        IPropertyOrParameterSymbol property,
         (INamedTypeSymbol key, INamedTypeSymbol value)? tuple)
     {
         var className = classSymbol.BuilderClassName;
@@ -29,9 +30,9 @@ internal partial class FluentBuilderClassesGenerator
         var cast = property.Type.TypeKind == TypeKind.Interface ? "" : $"({property.Type}) ";
 
         return new StringBuilder()
-            .AppendLine($"        public {className} With{property.Name}(Action<{_context.AssemblyName}.FluentBuilder.{dictionaryBuilderName}> action, bool useObjectInitializer = true) => With{property.Name}(() =>")
+            .AppendLine($"        public {className} With{property.Name}(Action<{_compilationHelper.AssemblyName}.FluentBuilder.{dictionaryBuilderName}> action, bool useObjectInitializer = true) => With{property.Name}(() =>")
             .AppendLine("        {")
-            .AppendLine($"            var builder = new {_context.AssemblyName}.FluentBuilder.{dictionaryBuilderName}();")
+            .AppendLine($"            var builder = new {_compilationHelper.AssemblyName}.FluentBuilder.{dictionaryBuilderName}();")
             .AppendLine("            action(builder);")
             .AppendLine($"            return {cast}builder.Build(useObjectInitializer);")
             .AppendLine("        });");