+++ src/FluentBuilderGenerator/SyntaxReceiver/IAutoGenerateBuilderSyntaxReceiver.cs
@@ -1,9 +1,0 @@
-using FluentBuilderGenerator.Models;
-using Microsoft.CodeAnalysis;
-
-namespace FluentBuilderGenerator.SyntaxReceiver;
-
-internal interface IAutoGenerateBuilderSyntaxReceiver : ISyntaxContextReceiver
-{
-    public IList<FluentData> CandidateFluentDataItems { get; }
-}