+++ tests/FluentBuilderGeneratorTests/DTO/ClassWithPrimaryConstructor.cs
@@ -1,0 +1,9 @@
+namespace FluentBuilderGeneratorTests.DTO
+{
+    public class ClassWithPrimaryConstructor(string test, int num)
+    {
+        public string Normal { get; set; } = string.Empty;
+
+        public string Data { get; } = test + num;
+    }
+}